models

+++ students/models.py
@@ -10,8 +10,8 @@ class Course(models.Model):
 
     department = models.ForeignKey(Department, on_delete=models.CASCADE, related_name='courses',)
 
-    semester = models.PositiveIntegerField(default=1)
-    credits = models.PositiveIntegerField(default=1)
+    semester = models.PositiveSmallIntegerField(default=1)
+    credits = models.PositiveSmallIntegerField(default=1)
    
     def __str__(self):
         return f"{self.first_name} {self.last_name}"
